pyproject.toml
[project]
name = "autostack"
version = "0.1.0"
description = "Autonomous multi-agent software development system"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AutoStack Team" }
]
keywords = ["ai", "agents", "langgraph", "automation", "hackathon"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    # LangGraph and LLM
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-core>=0.1.10",
    "langchain-groq>=0.3.0",
    "langchain-openai>=0.3.0",
    # Vector Store (thin client for Chroma Cloud)
    "chromadb-client>=0.5.0",
    # Azure OpenAI Embeddings for Chroma
    "openai>=1.0.0",
    # HTTP Client
    "httpx>=0.26.0",
    "httpx-retry>=0.1.0",
    "aiohttp>=3.9.1",
    # Notifications
    "apprise>=1.9.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    # Security
    "cryptography>=43.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    # Research (Tavily AI Search)
    "tavily-python>=0.5.0",
    # RepoMapper (Code structure analysis)
    "networkx>=2.6.3",
    "diskcache>=5.6.0",
    "grep-ast>=0.3.0",
    "tree-sitter>=0.20.0",
    "tiktoken>=0.5.0",
    # Infrastructure specific
    "checkov>=3.2.0",
    # LangGraph Postgres checkpointing
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.2.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    
    # Code Quality
    "black>=24.1.1",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.14",
]

[project.scripts]
autostack = "api.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "agents", "models", "services", "scripts"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "chromadb_client.*",
    "langgraph.*",
    "langchain.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["api", "agents", "models", "services"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
